requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
uvicorn>=0.32.1
pymysql>=1.1.0
sqlalchemy>=2.0.0
fastapi>=0.104.1
pydantic>=2.5.0
pandas>=2.0.0
# MCP SDK se instalará desde GitHub en el Dockerfile si está disponible